Refactor data collators classes to separate files for readability
